requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
matplotlib
numpy
pandas
pyyaml
scikit-learn
